requirements.txt
amqp==2.5.2
astroid==2.3.2
billiard==3.6.1.0
boto3==1.10.48
botocore==1.13.48
celery==4.4.0
certifi==2019.9.11
chardet==3.0.4
cryptography==2.8
Django==2.2.6
django-babeljs==0.2
django-bcrypt==0.9.2
django-celery-results==1.1.2
django-dynamic-fixture==2.0.0
django-elasticsearch-dsl==6.4.1
django-extensions==2.2.5
django-jsonfield==1.3.1
django-redis==4.10.0
django-simple-history==2.8.0
django-suit==0.2.28
djangorestframework==3.10.3
docutils==0.15.2
elasticsearch==6.4.0
elasticsearch-dsl==6.4.0
idna==2.8
importlib-metadata==1.3.0
isort==4.3.21
jmespath==0.9.4
kombu==4.6.7
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.0.2
mysqlclient==1.4.6
phonenumbers==8.10.21
py-bcrypt==0.4
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.3
pylint-django==2.0.11
pylint-plugin-utils==0.6
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
vine==1.3.0
wrapt==1.11.2
zipp==0.6.0
gevent
uwsgi
newrelic
sentry-sdk==0.14.4
redis==3.0.1
redis-py-cluster==2.0.0
kafka-python==2.0.1
JSON-log-formatter==0.2.0
django-log-request-id==1.6.0
django-requestlogs==0.2.1
elastic-apm
django-cors-headers==3.1.1
sendgrid==6.9.4
exponent-server-sdk==2.0.0
twilio